# semana13/api_role_sweets/app/static/swagger.json
{
   "openapi": "3.0.1",
   "info": {
      "title": "Dulceria API",
      "version": "1.0.0"
   },
   "paths": {
      "/api/sweets": {
         "get": {
            "summary": "Obtiene la lista de todos los Dulces",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Lista de Dulces",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/Sweet"
                           }
                        }
                     }
                  }
               }
            }
         },
         "post": {
            "summary": "Crea un nuevo Dulce",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/Sweet"
                     }
                  }
               }
            },
            "responses": {
               "201": {
                  "description": "Dulce creado",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Sweet"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/sweets/{id}": {
         "get": {
            "summary": "Obtiene un Dulce específico por su ID",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Detalles del Dulce",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Sweet"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Dulce no encontrado"
               }
            }
         },
         "put": {
            "summary": "Actualiza un Dulce existente por su ID",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/Sweet"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Dulce actualizado",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Sweet"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Dulce no encontrado"
               }
            }
         },
         "delete": {
            "summary": "Elimina un Dulce existente por su ID",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "204": {
                  "description": "Dulce eliminado"
               },
               "404": {
                  "description": "Dulce no encontrado"
               }
            }
         }
      },
      "/api/users": {
         "get": {
            "summary": "Obtiene la lista de todos los Usuarios",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Lista de Usuarios",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/User"
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/users/{id}": {
         "get": {
            "summary": "Obtiene un Usuario específico por su ID",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Detalles del Usuario",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/User"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Usuario no encontrado"
               }
            }
         },
         "put": {
            "summary": "Actualiza un Usuario existente por su ID",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/User"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Usuario actualizado",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/User"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Usuario no encontrado"
               }
            }
         },
         "delete": {
            "summary": "Elimina un Usuario existente por su ID",
            "security": [
               {
                  "JWTAuth": []
               }
            ],
            "parameters": [
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "204": {
                  "description": "Dulce eliminado"
               },
               "404": {
                  "description": "Dulce no encontrado"
               }
            }
         }
      },
      "/api/register": {
         "post": {
            "summary": "Registra un nuevo usuario",
            "requestBody": {
               "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/User"
                        }
                     }
               }
            },
            "responses": {
               "201": {
                     "description": "Usuario creado"
               },
               "400": {
                     "description": "Solicitud incorrecta"
               }
            }
         }
      },
      "/api/login": {
         "post": {
            "summary": "Inicia sesión con un usuario existente",
            "requestBody": {
               "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Login"
                        }
                     }
               }
            },
            "responses": {
               "200": {
                     "description": "Inicio de sesión exitoso",
                     "content": {
                        "application/json": {
                           "schema": {
                                 "type": "object",
                                 "properties": {
                                    "access_token": {
                                       "type": "string",
                                       "description": "Token de acceso JWT para el usuario"
                                    }
                                 }
                           }
                        }
                     }
               },
               "401": {
                     "description": "Credenciales inválidas"
               }
            }
         }
      }
   },
   
   "components": {
      "securitySchemes": {
         "JWTAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "schemas": {
         "Sweet": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "integer",
                  "readOnly": true
               },
               "brand": {
                  "type": "string"
               },
               "weight": {
                  "type": "number"
               },
               "flavor": {
                  "type": "string"
               },
               "origin": {
                  "type": "string"
               }
            },
            "required": [
               "brand",
               "weight",
               "flavor",
               "origin"
            ]
         },
         "User": {
            "type": "object",
            "properties": {
               "username": {
                  "type": "string"
               },
               "password": {
                  "type": "string"
               },
               "role": {
                  "type": "array",
                  "items": {
                     "type": "string",
                     "enum": [
                        "admin",
                        "user"
                     ]
                  }
               }
            },
            "required": [
               "username",
               "password",
               "role"
            ]
         },
         "Login": {
            "type": "object",
            "properties": {
               "username": {
                  "type": "string"
               },
               "password": {
                  "type": "string"
               }
            },
            "required": [
               "username",
               "password"
            ]
         }
      }
   }
}